names fixed, code refactored
browserwindow moved to utils, tray and browserwindow were separated, electronComponents renamed.
